Fixed an error with vector_enlarge which was due to incorrect reallocation

+++ vector/src/vector.h
@@ -25,7 +25,7 @@ typedef struct
     } Vector;
 
 Vector *vector_open(void);
-int vector_close(Vector **ppVector) __attribute__((nonnull (1)));
+size_t vector_close(Vector **ppVector) __attribute__((nonnull (1)));
 int vector_add(Vector *pVector, void *item) __attribute__((nonnull (1, 2)));
 int vector_removeAt(Vector *pVector, size_t index) __attribute__((nonnull (1)));
 void *vector_elementAt(Vector *pVector, size_t index) __attribute__((nonnull (1)));
